Organized webRTC commands
