+++ lib/services/mirv_api.dart
@@ -13,6 +13,9 @@ class MirvApi {
   String secondRover = "two";
   String thirdRover = "three";
   String _authToken = "auth token";
+  final String ipAdress =
+      // 'http://172.250.250.44:8000';
+      'https://mirvcloudapi.azurewebsites.net';
 
   BehaviorSubject<RoverMetrics> periodicMetricUpdates =
       new BehaviorSubject<RoverMetrics>();
@@ -23,8 +26,7 @@ class MirvApi {
   }
 
   Future<RoverMetrics> getRoverMetrics(String roverID) async {
-    var response =
-        await http.get(Uri.parse("http://44.202.152.178:8000/rovers/$roverID"));
+    var response = await http.get(Uri.parse("$ipAdress/rovers/$roverID"));
     String lol =
         '{"roverId":"rover1","state":"docked","status":"available","battery":22,"health":{"electronics":"healthy","drivetrain":"unavailable","intake":"healthy","sensors":"healthy","garage":"degraded","power":"unavailable","general":"degraded"},"telemetry":{"location":{"long":-104.969454,"lat":40.474101},"heading":149.68,"speed":12.62}}';
 
@@ -34,8 +36,7 @@ class MirvApi {
 
   Future<List<RoverSummary>> getRovers() async {
     List<RoverSummary> rovers;
-    var response =
-        await http.get(Uri.parse("http://44.202.152.178:8000/rovers"));
+    var response = await http.get(Uri.parse("$ipAdress/rovers"));
     rovers = (json.decode(response.body) as List)
         .map((i) => RoverSummary.fromJson(i))
         .toList()
